refactor(addcard): divide addcard functions with eventlistener
