+++ index.html
@@ -44,7 +44,7 @@
           </div>
         </li>
       </ol>
-      <ol class="column">
+      <ol class="column" data-column-key="2">
         <li class="task-list">
           <div class="column-title">진행중인 일</div>
           <div class="column-count">카운트</div>
@@ -52,7 +52,7 @@
           <div class="edit-btn"> x</div>
         </li>
       </ol>
-      <ol class="column">
+      <ol class="column" data-column-key="3">
         <li class="task-list">
           <div class="column-title">완료한 일</div>
           <div class="column-count">카운트</div>
